requirement.txt
qdrant_client==1.14.3
fastapi==0.115.9
sentence-transformers==4.0.1
pydantic==2.10.3
langchain_core==0.3.22
langchain_community==0.3.10
langchain_text_splitters==0.3.2
docx2txt==0.9
google-generativeai==0.8.5